update the port field in the DemoController to use UShort type

+++ doordeck-sdk/src/androidMain/kotlin/com/doordeck/multiplatform/sdk/model/data/FusionOperations.android.kt
@@ -63,7 +63,7 @@ object FusionOperations {
     ) : LockController
 
     data class DemoController(
-        val port: Int = 8080
+        val port: UShort = 8080u
     ) : LockController
     
     data class GallagherController(
